TS => MJS: text (top module)
Migrate fjs/text/module.f.ts to module.f.mjs under the stage-1
TypeScript-to-mjs migration, folded into the same PR since it became a
fresh dependency leaf once text/utf8 migrated. All other runtime deps
(bit_vec, list, utf16, nullable) were already .f.mjs.

proof.f.ts stays TypeScript: it still imports the unmigrated
types/string module; only the module import specifier is updated.

Update the 21 remaining .f.ts importers across the repo, found via an
exhaustive resolved-path scan rather than a literal-substring grep --
the substring approach missed fjs/text/sgr/module.f.ts's `../module.f.ts`
(no "text" segment in the specifier itself) on the first pass, the same
class of miss as the earlier crypto/secp/crypto/sign incident. Also
update three docs (two path-only, one re-anchored line-numbered quote
in vec-to-code-point-pipeline.md).
