+++ fjs/ci/proof.f.ts
@@ -2,7 +2,7 @@ import { ci, main } from './module.f.ts'
 import { functionalscript, node } from './config/module.f.mjs'
 import { nodeNixJobs } from './node/module.f.ts'
 import { coverageInclude } from './deno/module.f.ts'
-import { utf8, utf8ToString } from '../text/module.f.ts'
+import { utf8, utf8ToString } from '../text/module.f.mjs'
 import { empty as emptyVec } from '../types/bit_vec/module.f.mjs'
 import { type MetaStep, type Os, test, ubuntu, type GitHubAction, parseGitHubAction } from './common/module.f.ts'
 import { assert, assertEq } from '../asserts/module.f.mjs'
