+++ fjs/crypto/hmac/proof.f.ts
@@ -1,5 +1,5 @@
 import { assertEq } from '../../asserts/module.f.mjs'
-import { utf8 } from '../../text/module.f.ts'
+import { utf8 } from '../../text/module.f.mjs'
 import { uint, vec } from '../../types/bit_vec/module.f.mjs'
 import { sha256, sha384, sha512 } from '../sha2/module.f.mjs'
 import { hmac } from './module.f.ts'
