+++ fjs/crypto/pow/proof.f.ts
@@ -1,4 +1,4 @@
-import { utf8 } from '../../text/module.f.ts'
+import { utf8 } from '../../text/module.f.mjs'
 import { empty, uint } from '../../types/bit_vec/module.f.mjs'
 import { computeSync, sha224, sha256 } from '../sha2/module.f.mjs'
 import { bitcoinPow, genesisNBits, genesisTarget, pow, sha256Pow, targetFromNBits } from './module.f.mjs'
