+++ fjs/djs/transpiler/proof.f.ts
@@ -2,7 +2,7 @@ import { sort } from '../../types/object/module.f.ts'
 import { transpile } from './module.f.ts'
 import { stringifyAsTree } from '../serializer/module.f.ts'
 import { virtual, emptyState, type Dir } from '../../effects/node/virtual/module.f.ts'
-import { utf8 } from '../../text/module.f.ts'
+import { utf8 } from '../../text/module.f.mjs'
 import { assert, assertEq } from '../../asserts/module.f.mjs'
 
 const run = (root: Dir) => (path: string) => {
