+++ fjs/effects/node/virtual/module.f.ts
@@ -5,7 +5,7 @@
  */
 import { todo } from '../../../asserts/module.f.mjs'
 import { isProperPrefix, join, parse } from '../../../path/module.f.ts'
-import { utf8ToString } from '../../../text/module.f.ts'
+import { utf8ToString } from '../../../text/module.f.mjs'
 import { empty, length, maxLengthBytes, msb, vec, type Vec } from '../../../types/bit_vec/module.f.mjs'
 import { error, ok } from '../../../types/result/module.f.mjs'
 import { run, type MemOperationMap, type RunInstance } from '../../mock/module.f.ts'
