+++ fjs/mcp/cas/module.f.ts
@@ -107,7 +107,7 @@ import { pure, step, type Effect } from '../../effects/module.f.ts'
 import { type MemOp } from '../../effects/memory/module.f.ts'
 import { cBase32ToVec, vecToCBase32 } from '../../basen/cbase32/module.f.mjs'
 import { decode as base64Decode, encode as base64Encode } from '../../basen/base64/module.f.mjs'
-import { tryUtf8 } from '../../text/module.f.ts'
+import { tryUtf8 } from '../../text/module.f.mjs'
 import { detectStream } from '../../media/type/module.f.ts'
 import { detect } from '../../media/module.f.ts'
 import { revisionDialect } from '../../media/revision/module.f.ts'
