+++ fjs/mcp/proof.f.ts
@@ -8,7 +8,7 @@ import type { Response } from '../protocol/json_rpc/module.f.ts'
 import { msb, u8ListToVec, vec8, repeat, length, type Vec, maxLengthBytes } from '../types/bit_vec/module.f.mjs'
 import { vecToCBase32 } from '../basen/cbase32/module.f.mjs'
 import { encode as base64Encode } from '../basen/base64/module.f.mjs'
-import { utf8 } from '../text/module.f.ts'
+import { utf8 } from '../text/module.f.mjs'
 import { fileCas, type FileCasOperation } from '../cas/module.f.ts'
 import { dialect as revisionDialect, mediaType as revisionMediaType } from '../media/revision/module.f.ts'
 import { sha256 } from '../crypto/sha2/module.f.mjs'
