+++ fjs/media/html/module.f.ts
@@ -11,7 +11,7 @@ import { compose } from '../../types/function/module.f.mjs'
 import { stringToList } from '../../text/utf16/module.f.mjs'
 import { includes } from '../../types/array/module.f.mjs'
 import { type Vec } from '../../types/bit_vec/module.f.mjs'
-import { utf8 } from '../../text/module.f.ts'
+import { utf8 } from '../../text/module.f.mjs'
 import { quotationMark, ampersand, lessThanSign, greaterThanSign } from '../../text/ascii/module.f.mjs'
 
 const { fromCharCode } = String
