+++ fjs/protocol/mcp/stdio/proof.f.ts
@@ -3,7 +3,7 @@ import { pure, type Effect } from '../../../effects/module.f.ts'
 import { emptyState, virtual, type State } from '../../../effects/node/virtual/module.f.ts'
 import type { Unknown } from '../../../media/json/module.f.ts'
 import { stringify } from '../../../media/json/module.f.ts'
-import { utf8 } from '../../../text/module.f.ts'
+import { utf8 } from '../../../text/module.f.mjs'
 import { fromVec } from '../../../types/uint8array/module.f.ts'
 import { maxLengthBytes } from '../../../types/bit_vec/module.f.mjs'
 import { sort } from '../../../types/object/module.f.ts'
