+++ fjs/sul/id/module.f.ts
@@ -17,7 +17,7 @@ import {
     type Vec
 } from '../../types/bit_vec/module.f.mjs'
 import { assertEq } from '../../asserts/module.f.mjs'
-import { utf8 } from '../../text/module.f.ts'
+import { utf8 } from '../../text/module.f.mjs'
 import { secp256r1, type Point2D } from '../../crypto/secp/module.f.mjs'
 import { base32, type V8 } from '../../crypto/sha2/module.f.mjs'
 import { literal3ToVec } from '../level/literal/module.f.ts'
