+++ fjs/text/proof.f.ts
@@ -1,5 +1,5 @@
 import { assertEq, assertNotNullish } from '../asserts/module.f.mjs'
-import { flat, utf8, utf8ToString, tryUtf8, type Block } from './module.f.ts'
+import { flat, utf8, utf8ToString, tryUtf8, type Block } from './module.f.mjs'
 import { join } from '../types/string/module.f.ts'
 import { empty, maxLengthBytes } from '../types/bit_vec/module.f.mjs'
