+++ fjs/types/uint8array/module.f.ts
@@ -10,7 +10,7 @@
  * @module
  */
 import { assert, assertNotNullish } from '../../asserts/module.f.mjs'
-import { utf8, utf8ToString } from '../../text/module.f.ts'
+import { utf8, utf8ToString } from '../../text/module.f.mjs'
 import { maxLengthBytes, msb, tryU8ListToVec, u8List, u8ListToVec, type Vec } from '../bit_vec/module.f.mjs'
 import { compose } from '../function/module.f.mjs'
 import { flat, fromArrayLike, iterable, map, type List } from '../list/module.f.mjs'
